Some finishing updates and fixing typos.
